requirements.txt
amqp==5.1.0
anyjson==0.3.3
asgiref==3.5.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.12
django-celery-beat==2.2.1
django-filter==21.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.11.3
kombu==5.2.4
Markdown==3.3.6
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
redis==3.5.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
zipp==3.7.0
